pyproject.toml
[project]
name = "sme_business_coach"
version = "0.1.0"
description = "sme-business-coach using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.152.0,<1.0.0",
    "streamlit>=1.35.0",
    "python-dotenv>=1.0.1"
]

[project.scripts]
sme_business_coach = "sme_business_coach.main:run"
run_crew = "sme_business_coach.main:run"
train = "sme_business_coach.main:train"
replay = "sme_business_coach.main:replay"
test = "sme_business_coach.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
